WIP: Refactor test support
